refactor "cancel request"
