+++ src/main/java/org/folio/innreach/controller/d2ir/InnReachCirculationController.java
@@ -7,11 +7,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import org.folio.innreach.domain.service.CirculationService;
-import org.folio.innreach.dto.CirculationRequestDTO;
+import org.folio.innreach.dto.CancelRequestDTO;
 import org.folio.innreach.dto.InnReachResponseDTO;
 import org.folio.innreach.dto.ItemShippedDTO;
 import org.folio.innreach.dto.PatronHoldDTO;
@@ -26,15 +25,6 @@ public class InnReachCirculationController implements InnReachCirculationApi {
 
   private final CirculationService circulationService;
 
-  @Override
-  @RequestMapping(value = "/{circulationOperationName}/{trackingId}/{centralCode}", method = {RequestMethod.POST, RequestMethod.PUT})
-  public ResponseEntity<InnReachResponseDTO> processCirculationRequest(@PathVariable String trackingId,
-                                                                       @PathVariable String centralCode,
-                                                                       @PathVariable String circulationOperationName,
-                                                                       CirculationRequestDTO circulationRequest) {
-    var innReachResponse = circulationService.processCirculationRequest(trackingId, centralCode, circulationOperationName, circulationRequest);
-    return ResponseEntity.ok(innReachResponse);
-  }
 
   @Override
   @PostMapping("/patronHold/{trackingId}/{centralCode}")
@@ -50,6 +40,15 @@ public class InnReachCirculationController implements InnReachCirculationApi {
   public ResponseEntity<InnReachResponseDTO> itemShipped(@PathVariable String trackingId,
       @PathVariable String centralCode, ItemShippedDTO itemShipped) {
     var innReachResponse = circulationService.trackShippedItem(trackingId, centralCode, itemShipped);
+
+    return ResponseEntity.ok(innReachResponse);
+  }
+
+  @Override
+  @PutMapping("/cancelrequest/{trackingId}/{centralCode}")
+  public ResponseEntity<InnReachResponseDTO> cancelRequest(@PathVariable String trackingId,
+      @PathVariable String centralCode, CancelRequestDTO cancelRequest) {
+    var innReachResponse = circulationService.cancelTransaction(trackingId, centralCode, cancelRequest);
 
     return ResponseEntity.ok(innReachResponse);
   }
