--- src/main/java/org/folio/innreach/domain/CirculationOperation.java
@@ -1,16 +1,0 @@
-package org.folio.innreach.domain;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum CirculationOperation {
-
-  CANCEL_ITEM_HOLD ("cancel_itemhold"),
-  CANCEL_PATRON_HOLD("cancelrequest");
-
-  private final String operationName;
-
-}
