+++ src/main/java/org/folio/innreach/domain/service/CirculationService.java
@@ -1,6 +1,6 @@
 package org.folio.innreach.domain.service;
 
-import org.folio.innreach.dto.CirculationRequestDTO;
+import org.folio.innreach.dto.CancelRequestDTO;
 import org.folio.innreach.dto.InnReachResponseDTO;
 import org.folio.innreach.dto.ItemShippedDTO;
 import org.folio.innreach.dto.PatronHoldDTO;
@@ -8,12 +8,11 @@ import org.folio.innreach.dto.TransferRequestDTO;
 
 public interface CirculationService {
 
-  InnReachResponseDTO processCirculationRequest(String trackingId, String centralCode, String circulationOperationName, CirculationRequestDTO circulationRequest);
-
   InnReachResponseDTO initiatePatronHold(String trackingId, String centralCode, PatronHoldDTO patronHold);
 
   InnReachResponseDTO trackShippedItem(String trackingId, String centralCode, ItemShippedDTO itemShipped);
 
-  InnReachResponseDTO transferItem(String trackingId, String centralCode, TransferRequestDTO transferRequest);
+  InnReachResponseDTO cancelTransaction(String trackingId, String centralCode, CancelRequestDTO cancelRequest);
 
+  InnReachResponseDTO transferItem(String trackingId, String centralCode, TransferRequestDTO transferRequest);
 }
